backend/requirements.txt
# Base
pillow== 7.2.0
markdown== 3.3.1
pytz>=2020.1

# DataBase Conection
psycopg2-binary== 2.8.6

# Enviroment
django-environ== 0.4.5

# Django
Django>= 3.1.2

# Django REST Framework
djangorestframework== 3.12.1
django-filter== 2.4.0
django-cleanup== 5.1.0
django-oauth-toolkit>= 1.3.2

# JWT
pyjwt>= 1.7.1
djangorestframework-simplejwt >= 4.4.0

# Password security
argon2-cffi>= 20.1.0

# Static files
whitenoise== 5.2.0

# Conection frontend
django-cors-headers == 3.5.0
